app/requirements.txt
python-dotenv
PyYAML
sentence_transformers 
einops # If using PEFT fine-tuned model
bitsandbytes 
accelerate  
transformers[torch] 
xformers # If using PEFT fine-tuned model
datasets 
trl # If using PEFT fine-tuned model
peft # If using PEFT fine-tuned model
torch 
snowflake-snowpark-python[pandas]
pandas
weaviate-client==3.21.0
vllm==0.2.2
typing_extensions==4.5.0
pydantic==1.10.13
fschat==0.2.33
openai==1.3.4
streamlit
langchain==0.0.341 # Needs to be at end
numexpr # Needed for calculator LLM